+++ init/initrd/dracut/info.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+man dracut.bootup | cat